First commit mocks code
